requirements.txt
scikit-learn==0.20
absl-py==0.8.1
alabaster==0.7.12
appnope==0.1.0
astor==0.8.0
attrs==19.3.0
awscli==1.16.161
Babel==2.6.0
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.0
boto3==1.10.20
botocore==1.13.20
cached-property==1.5.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
colorama==0.3.9
cryptography==2.8
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
dtw==1.3.3
entrypoints==0.3
gast==0.2.2
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.7
imagesize==1.1.0
importlib-metadata==0.23
ipdb==0.12.2
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10
jmespath==0.9.4
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.16.2
opt-einsum==3.1.0
packaging==19.0
pandas==0.24.1
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.0
protobuf3-to-dict==0.1.5
ptyprocess==0.6.0
pyasn1==0.4.5
pycparser==2.19
Pygments==2.3.1
PyNaCl==1.3.0
pyparsing==2.3.1
pyrsistent==0.15.5
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
pyzmq==18.1.1
qtconsole==4.5.5
requests==2.20.1
rsa==3.4.2
s3transfer==0.2.0
sagemaker==1.25.0
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==1.8.4
sphinxcontrib-websupport==1.1.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
texttable==1.6.2
tornado==6.0.3
traitlets==4.3.3
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0

